requirements.txt
gdpc==7.1.0
networkx==3.3
numpy==1.26.4
Pillow==10.3.0
pygame==2.5.2
scipy==1.13.1
skan==0.11.1
skimage==0.0
